+++ JavaScriptEngineSwitcher.Jint/Properties/AssemblyInfo.cs
@@ -6,12 +6,12 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("JavaScript Engine Switcher for .Net: Jint")]
-[assembly: AssemblyCopyright("Copyright © Andrey Taritsyn 2014")]
+[assembly: AssemblyCopyright("Copyright © 2013-2015 Andrey Taritsyn")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
 [assembly: ComVisible(false)]
 [assembly: Guid("4b51319e-014f-4731-8a80-15b5c58f3a72")]
 
-[assembly: AssemblyVersion("1.2.1.0")]
-[assembly: AssemblyFileVersion("1.2.1.0")]
+[assembly: AssemblyVersion("1.2.3.0")]
+[assembly: AssemblyFileVersion("1.2.3.0")]