+++ JavaScriptEngineSwitcher.Tests/Properties/AssemblyInfo.cs
@@ -6,12 +6,12 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyConfiguration("")]
 [assembly: AssemblyCompany("")]
 [assembly: AssemblyProduct("JavaScript Engine Switcher for .Net: Tests")]
-[assembly: AssemblyCopyright("Copyright © 2012-2015 Andrey Taritsyn")]
+[assembly: AssemblyCopyright("Copyright © 2013-2015 Andrey Taritsyn")]
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 
 [assembly: ComVisible(false)]
 [assembly: Guid("4cfa9f30-6718-4b7e-8329-7602278b8e31")]
 
-[assembly: AssemblyVersion("1.2.2.0")]
-[assembly: AssemblyFileVersion("1.2.2.0")]
+[assembly: AssemblyVersion("1.2.3.0")]
+[assembly: AssemblyFileVersion("1.2.3.0")]